--- a/src/python/pants/core/goals/run_test.py
+++ b/src/python/pants/core/goals/run_test.py
@@ -36,7 +36,7 @@ from pants.engine.target import (
     TargetRootsToFieldSets,
     TargetRootsToFieldSetsRequest,
 )
-from pants.engine.unions import UnionMembership
+from pants.engine.unions import UnionMembership, UnionRule
 from pants.option.global_options import GlobalOptions, KeepSandboxes
 from pants.testutil.option_util import create_goal_subsystem, create_subsystem
 from pants.testutil.rule_runner import (
@@ -138,10 +138,10 @@ def single_target_run(
                     mock=rule_runner.run_interactive_process,
                 ),
             ],
-            union_membership=UnionMembership(
+            union_membership=UnionMembership.from_rules(
                 {
-                    RunFieldSet: [TestRunFieldSet],
-                    RunDebugAdapterRequest: [TestRunFieldSet],
+                    UnionRule(RunFieldSet, TestRunFieldSet),
+                    UnionRule(RunDebugAdapterRequest, TestRunFieldSet),
                 },
             ),
         )