I expect it to be a net-nothing performance wise, but I think `UnionMembership.union_rules` need to be available rust-side for clean rust-migration of rules & targets as it sees plenty of use.

Before:

```console
hyperfine -w 2 -r 10 './pants --no-pantsd dependencies ::'                                                        
Benchmark 1: ./pants --no-pantsd dependencies ::
  Time (mean ± σ):     25.734 s ±  1.620 s    [User: 26.290 s, System: 7.470 s]
  Range (min … max):   24.437 s … 29.825 s    10 runs
```
After:
```console
hyperfine -w 2 -r 10 './pants --no-pantsd dependencies ::'
Benchmark 1: ./pants --no-pantsd dependencies ::
  Time (mean ± σ):     25.728 s ±  1.858 s    [User: 26.209 s, System: 7.553 s]
  Range (min … max):   23.437 s … 29.024 s    10 runs
```